tests/texdown: cast pointer to uintptr_t
This avoids a compiler-warning on MSVC.

+++ src/tests/texdown.c
@@ -184,7 +184,7 @@ MeasureDownloadRate(void)
 
    printf("alloc %p\n", texImage);
 
-   for (i = 1; !(((unsigned long)texImage) & i); i<<=1)
+   for (i = 1; !(((uintptr_t)texImage) & i); i<<=1)
       ;
    printf("texture image alignment: %d bytes (%p)\n", i, texImage);
